feat(parser): integrate dice parser and add parser tests (T5)
