+++ server/package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "dev": "ts-node-dev --respawn --transpile-only src/index.ts",
     "start": "node dist/index.js",
-    "migrate": "ts-node src/db/runMigrations.ts"
+    "migrate": "ts-node src/db/runMigrations.ts",
+    "test:parser": "ts-node src/test-parser.ts"
   },
   "dependencies": {
     "express": "^4.18.2",
